Considered Model 2 regression
Added a section of notes and a bit of code to explore the option of using a model II regression (see Sokal & Rohl; Legendre & Legendre), given that my 'X' values are not 'fixed', and there is error associated with both the response and continuous explanatory variables.
The notes explain why I am abandoning this approach, but briefly: 
It might work in theory, but I'm not convinced the data really satisfies all the assumptions of even a model 2 regression, either, or that the relationship is simply "linear", and there are no ready tools to easily do what I want to do (compare the relationships across treatment groups).
It's probably just easier to go with stardard OLS regression, take the results with a grain of salt and accept that the regression slopes are likely underestimated.

Also added a trivial function 'gmean' to calculate the geometric mean
+ in 'jaw.misc_functions.R'


Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>