ARA~cb Code Reorganization for Modularity
Separate scripts to allow modularity, and make it easier to try alternative methods (and abandon them if fruitless).
- Master Control script: source()s selected others in order, for Final Analysis.
- Load & process data; configure analysis options
- Data Exploration
- Modelling
- model II regression (abandoned)
- Meta-Community scale analysis (started, but not encouraging)

The number of options and alternatives was getting too unwieldy, and I was getting tired of scrolling through big chunks of script.
This should make it easier to make copies of scripts for variations, then use the master control script to decide which ones to use in the Final Analysis.


Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>