Tinkering to improve Frag icon labels (experimental)
It occurs to me that icons alone may be unclear, especially if there is no direct connection between the plot labels and the words used in the text.  
Currently, the icons are directly linked to numbers in a diagram, with the numbers linked to the words in the legend.  Too many steps.

- overlay numbers on top / inside of icons on graph?
- add a legend with text labels beside icons?  or do this once somewhere else?
  - this could be added directly in R, but it might actually make more sense to do this in LaTeX.

Also added `SECCicons()` function to auto-load all icons in `/save/` as a named list.



Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>