Separate analyses for N-availability: with and without exposure time
Files with "-time" refer to analysis without taking into account
different exposure times / duration of ion resin capsules.
This demonstrates that there is no difference in total amounts,
and therefore the total adsorption was similar for capsules left out
only during the summer, as those left out for an entire year.
The implication is that most N is adsorbed during the summer.

Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>