Preliminary analysis of Moss Growth / Productivity
Using nested mixed models (lme)
multcomp for multiple comparisons (backup to measure true differences between treatment levels)
effects for graphical displays (95% confidence intervals: NOT corrected for multiple comparisons).

Separate analysis at each measurement time (separate variables)
Separate results joined at the end for ggplot graph.

There is a suggestion of an overall effect of time: more productivity at first measurement.
However, given that the first measurement spans an amount of time equal to both the second and third measurement periods *put together*, this is not surprising.
What is surprising is that the last two measurements are equal, and large compared to the first.
This suggests that the moss is growing just as much over the 9 months of winter, as the 3 months of summer.  It also suggests that productivity *increased* the second year...

I might try with combined growth measurements, or divide production by time,
to convert the values to rates.
Might be worth performing a proper repeated measures analysis through time,
but keep in mind that the time intervals are **NOT** equal.


Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>