README and GNU GPLv3 license
In preparation for sharing this project on GitHub, 
I've added the GNU GPLv3 (or later) to the README file, and a full copy of the license to `/doc`.
I'm not entirely sure this is the best license to use, but only because it requires
any modified versions to also use a free license (copy*left*).
This is great for preventing people from creating non-free products using my data,
but I don't want it to be incompatible with other free licenses.
Basically, I want to allow others to use the data or modify the source code, as long as I get credit for it :)

I'm hoping this will be ok, as R itself is released under GPLv2 (or later), and packages are released under a range of compatible free licenses:

http://www.gnu.org/licenses/
http://www.R-project.org/Licenses/



Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>