Fauna analyses now produce final graphs for 2-way interactions (that are significant)
The 3-way interaction between Fragmentation, Chamber, and Position, is NS, so it is skipped (graph is still produced, but not saved / exported).
I suspect there is simply too much noise, due to low replication for the 3-way interaction to be detected.
It actually makes little sense for both CxP, and FxP to be significant without CxFxP.

Basically, only Outer Isolated patches show detectable declines, although the general patterns of relative values is pretty consistent.



Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>