Fixed miscalculation of patch area (re-run CH1)
Yeah, I'm an idiot. I had been calculating patch area using the diameter, not the radius (*face slap*).
The calculation has been fixed in `SECC.functions.R`, and everywhere else I could find it.
Just to be safe, I blocked out many of the definitions of these constants, to allow the values in `SECC.functions.R` to be used instead.
I re-ran analyses for Chapter 1 just to be sure: no major changes or errors.
Thankfully, it appears that this value was never really used until now (in CH3).


Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>