Check for Block effects in Productivity model
They are statistically significant, although barely.
They may be marginal.
The point is that I don't care about variation across Blocks, that's expected.
I'm more worried about possible interaction, which appears significant, but not serious in terms of magnitude (the basic pattern is the same, it's just weaker in some blocks).