Check effect of "Time" (replicates) on Moss Growth
There is a significant interaction term between 'Time' and the usual Chamber x Frag,
But this is only significant in the second interval (grow12),
and seems to have more to do with smaller differences in t3 patches,
but with the same overall trend.
I'm hand-waving this "effect" away as spurious, given that there really shouldn't *be* an effect of replicate ID in this situation.


Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>