Some Model selection to verify Net Production (PD.bal) model
Accounting for heterogeneity across Frag treatments makes for a *slightly*
better fit (according to AIC< but not BIC), and renders Frag treatments NS.
(residuals look just as normal).
BUT, this throws errors with the `effects` package, and makes plotting pretty much impossible.
I'm calling the Frag effect (and interaction terms) NS, and using the regular model for plotting.

Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>