CH4: TAN ANOVA - fixed erroneous transformation, re-ran analyses
I discovered I had used the default log10(Y + 1) transformation, which was totally messing with the variance.  The +1 is unnecessary and problematic for all these small non-0 values.  I overrode the default with a custom logTAN column, which I used instead.

I also updated the final graphics code, and added a barplot of main Position effects, which are usually the only significant term (albeit apparently driven by what's happening in the Chambers).

Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>