CH4: Separate trial script using 'Climate' pseudo-factor
- Using the full Chamber:Position Interaction / 'Climate' takes a lot of explanatory power away from moisture.
  - moisture is confounded in the Chamber:Position interaction to begin with.

I have decided it is better to leave Position out, and just go with Chamber, to leave more for moisture to explain directly.
There is still the problem of no dry patches in the Ambient treatment, so I will avoid Chamber-interactions to avoid over-fitting and unrealistic predicted values in non-existent dry ambient patches.

Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>