CH4: Decomposition partial regression
The sign of the partial regression of Decomposition on Richness changes depending on whether H2o is included.
If I account for H2O, Richness is negative.
If I leave H2O out, Richness is positive.
Either way, the relationship is not significant.
I interpret this as moisture driving everything, and Richness doesn't explain much of what's left over.

Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>