Re-ran ALL analyses (CH1, CH2, CH3, CH4) with new factor level labels, and original contrasts.
That was painful figuring out the issue with contrasts.
The nested ANOVA script changed the settings for contrasts
	options(contrasts = c("contr.sum", "contr.poly"))
which is different from the default
	options(contrasts = c("contr.treatment", "contr.poly"))
Therefore, running scripts *after* the ANOVA scripts generated different model fits ans `summary()` output (including all the glmulti runs) :(
I added a line in `lib/init.R` to reset the options to the defaults, and now everything behaves as expected.
*sigh*



Signed-off-by: Jonathan A. Whiteley <jonathan.whiteley@mail.mcgill.ca>